requirements.txt
streamlit==1.52.2
requests==2.32.3
google-genai==1.56.0
google-cloud-firestore==2.22.0
pytest==9.0.2
black==25.12.0
